Move reset input into C file
