+++ Dark Samurai Warrior Reloaded/Dark Samurai Warrior Reloaded.vcxproj
@@ -130,7 +130,9 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="gui\gui.c" />
+    <ClCompile Include="input\input.c" />
     <ClCompile Include="main.c" />
+    <ClCompile Include="math.c" />
     <ClCompile Include="render.c" />
   </ItemGroup>
   <ItemGroup>
@@ -139,6 +141,7 @@
   <ItemGroup>
     <ClInclude Include="common.h" />
     <ClInclude Include="gui\gui.h" />
+    <ClInclude Include="input\input.h" />
     <ClInclude Include="math.h" />
     <ClInclude Include="render.h" />
   </ItemGroup>
