+++ Dark Samurai Warrior Reloaded/input/input.c
@@ -1,0 +1,9 @@
+#include "./input.h"
+
+void reset_input(Input *input) {
+  input->downEndedDown = 0;
+  input->leftEndedDown = 0;
+  input->rightEndedDown = 0;
+  input->upEndedDown = 0;
+  input->tabEndedDown = 0;
+}
